fix: 투표 전역 상태 초기화 로직 수정
